// SignalRClient-frontend/src/generated/SignalRClient.Api.ts
/* THIS (.ts) FILE IS GENERATED BY Tapper */
/* eslint-disable */
/* tslint:disable */

/** Transpiled from SignalRClient.Api.Message */
export type Message = {
    /** Transpiled from string */
    username: string;
    /** Transpiled from string */
    content: string;
    /** Transpiled from System.DateTime */
    timeStamp: (Date | string);
}